[ANDROSDK-1056] Move transaction and database adapter tests
