+++ core/src/test/java/org/hisp/dhis/android/core/arch/db/access/internal/SqLiteTransactionShould.java
@@ -26,10 +26,9 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.core.data.database;
+package org.hisp.dhis.android.core.arch.db.access.internal;
 
 import org.hisp.dhis.android.core.arch.db.access.DatabaseAdapter;
-import org.hisp.dhis.android.core.arch.db.access.internal.SqLiteTransaction;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
